#6 Combined separate CRUD services into single service
