+++ WebAPI/Controllers/AuthorsController.cs
@@ -7,31 +7,27 @@ using WebAPI.DTOs;
 using WebAPI.Interfaces;
 using Swashbuckle.AspNetCore.Annotations;
 using WebAPI.Models;
+using Microsoft.AspNetCore.Authorization;
 
 namespace WebAPI.Controllers
 {
+    //[Authorize(Roles = "Admin")]
     [Route("api/[controller]")]
     [ApiController]
     public class AuthorsController : ControllerBase
     {
-        private readonly ICreatorService<Author> _creator;
-        private readonly IGetterService<Author> _getter;
-        private readonly IUpdaterService<Author> _updater;
-        private readonly IDeleterService<Author> _deleter;
+        private readonly ICrudService<Author> _crud;
 
-        public AuthorsController(ICreatorService<Author> creator, IGetterService<Author> getter, IUpdaterService<Author> updater, IDeleterService<Author> deleter)
+        public AuthorsController(ICrudService<Author> crud)
         {
-            _creator = creator;
-            _getter = getter;
-            _updater = updater;
-            _deleter = deleter;
+            _crud = crud;
         }
 
         [SwaggerOperation(Summary = "Retrieves all book authors")]
         [HttpGet]
         public async Task<IActionResult> Index()
         {
-            var authors = await _getter.GetAll<AuthorResponseDto>();
+            var authors = await _crud.GetAll<AuthorResponseDto>();
             
             return Ok(authors);
         }
@@ -40,7 +36,7 @@ namespace WebAPI.Controllers
         [HttpGet("{id:int}")]
         public async Task<IActionResult> Show(int id)
         {
-            var author = await _getter.GetById<AuthorResponseDto>(id);
+            var author = await _crud.GetById<AuthorResponseDto>(id);
             
             return author == null ? NotFound() : Ok(author);
         }
@@ -48,7 +44,7 @@ namespace WebAPI.Controllers
         [HttpPost]
         public async Task<IActionResult> Create(AuthorRequestDto requestDto) 
         {
-            var author = await _creator.Create<AuthorResponseDto>(requestDto);
+            var author = await _crud.Create<AuthorResponseDto>(requestDto);
 
             return Created($"api/authors/{author.Id}", author);
         }
@@ -57,7 +53,7 @@ namespace WebAPI.Controllers
         [HttpPut("{id:int}")]
         public async Task<ActionResult> Update(int id, AuthorRequestDto requestDto)
         {
-            await _updater.Update(requestDto, id);
+            await _crud.Update(requestDto, id);
         
             return Ok("Resource updated");
         }
@@ -66,7 +62,7 @@ namespace WebAPI.Controllers
         [HttpDelete("{id:int}")]
         public async Task<ActionResult> Delete(int id)
         {
-            await _deleter.Delete(id);
+            await _crud.Delete(id);
           
             return Ok("Resource deleted");
         }
