+++ WebAPI/Installers/ServiceInstaller.cs
@@ -17,14 +17,8 @@ namespace WebAPI.Installers
         {
             services.AddScoped<ICategoryService, CategoryService>();
             services.AddSingleton<ILoggedUserProvider, LoggedUserProvider>();
-            services.AddScoped<ICreatorService<Author>, CreatorService<Author>>();
-            services.AddScoped<IGetterService<Author>, GetterService<Author>>();
-            services.AddScoped<IUpdaterService<Author>, UpdaterService<Author>>();
-            services.AddScoped<IDeleterService<Author>, DeleterService<Author>>();
-            services.AddScoped<ICreatorService<Book>, CreatorService<Book>>();
-            services.AddScoped<IGetterService<Book>, GetterService<Book>>();
-            services.AddScoped<IUpdaterService<Book>, UpdaterService<Book>>();
-            services.AddScoped<IDeleterService<Book>, DeleterService<Book>>();
+            services.AddScoped<ICrudService<Author>, CrudService<Author>>();
+            services.AddScoped<ICrudService<Book>, CrudService<Book>>();
         }
     }
 }
