--- WebAPI/Interfaces/ICreatorService.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using WebAPI.Common;
-using WebAPI.DTOs;
-
-namespace WebAPI.Interfaces
-{
-    public interface ICreatorService<T> where T : IDbModel
-    {
-        public Task<IReponseDto> Create<IReponseDto>(IRequestDto model);
-    }
-}
