+++ WebAPI/Interfaces/ICrudService.cs
@@ -3,13 +3,15 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using WebAPI.Common;
-using WebAPI.DTOs;
 
 namespace WebAPI.Interfaces
 {
-    public interface IGetterService<T> where T : IDbModel
+    public interface ICrudService<T> where T : IDbModel
     {
+        public Task<IReponseDto> Create<IReponseDto>(IRequestDto model);
         public Task<IEnumerable<IResponseDto>> GetAll<IResponseDto>();
         public Task<IResponseDto> GetById<IResponseDto>(int id);
+        public Task Update(IRequestDto updateModel, int id);
+        public Task Delete(int id);
     }
 }
