--- WebAPI/Interfaces/IDeleterService.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using WebAPI.Common;
-using WebAPI.Models;
-
-namespace WebAPI.Interfaces
-{
-    public interface IDeleterService<T> where T : IDbModel
-    {
-        public Task Delete(int id);
-    }
-}
